marketplace-scraper/requirements.txt
# Core requirements
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
google-cloud-firestore>=2.3.0
google-cloud-pubsub>=2.7.1
google-cloud-monitoring>=2.8.0
google-cloud-storage>=1.42.0
google-cloud-secret-manager>=2.8.0
firebase-admin>=5.0.3
cachetools>=4.2.4
python-dateutil>=2.8.2
urllib3>=1.26.6
requests>=2.26.0
async-timeout>=4.0.2
attrs>=21.2.0
multidict>=5.1.0
yarl>=1.6.3

# For browser actions and parsing
parsel>=1.6.0
selectolax>=0.3.12
pyppeteer>=1.0.2

# Data handling
pydantic>=1.8.2
jsonschema>=4.4.0

# For development and testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.1
pylint>=2.10.2
black>=21.8b0
flake8>=3.9.2
mypy>=0.910